# DO NOT SUBMIT [runtime,compiler] use custom ctype functions implementations