--- a/compiler/utils/string-utils.cpp
+++ b/compiler/utils/string-utils.cpp
@@ -4,18 +4,20 @@
 
 #include "compiler/utils/string-utils.h"
 
+#include "common/php-functions.h"
+
 // perform "blahBlah" to "blah_blah" translation
 std::string transform_to_snake_case(vk::string_view origin) noexcept {
   std::string name;
   name.reserve(static_cast<std::size_t>(origin.size() * 1.2));
 
   for (char c : origin) {
-    if (std::isupper(c)) {
+    if (php_isupper(c)) {
       if (!name.empty() && name.back() != '_') {
         name.append(1, '_');
       }
     }
-    name.append(1, std::tolower(c));
+    name.append(1, php_tolower(c));
   }
   return name;
 }
@@ -35,7 +37,7 @@ std::string transform_to_camel_case(vk::string_view origin) noexcept {
     char cur = origin[i];
     char next = origin[i + 1];
     if (cur == '_' && next) {
-      name.append(1, std::toupper(next));
+      name.append(1, php_toupper(next));
       ++i;
     } else {
       name.append(1, cur);