… common/php-functions.cpp