Functions from `<cctype>` are not inlined by `g++` or `clang++` even with `-O2` and `-O3`. This causes the major slowdown in multiple functions, `strtolower`, `strtoupper`, `strcasecmp`, `lcfirst,` `ucfirst` - just to name a few.

These custom functions should be used when php ctype functions will be added too.

This patch replaces all usages of `ctype.h` functions from both runtime and compiler.

Tests check that new functions return identical results for all possible inputs (values from 0 to 255, inclusive).

`strcasecmp` benchmark results for difference input string lengths:

	name               old time/op  new time/op  delta
	Strcasecmp::X32     344ns ± 2%    90ns ± 2%  -73.76%  (p=0.000 n=10+10)
	Strcasecmp::X256   2.74µs ± 1%  0.48µs ± 0%  -82.58%  (p=0.000 n=10+8)
	Strcasecmp::X1024  10.9µs ± 3%   1.8µs ± 0%  -83.12%  (p=0.000 n=10+8)
	[Geo mean]         2.17µs       0.43µs       -80.24%