@quasilyte 2022-09-26
Well. It's harder than it looks.

In some places, PHP functions use their own notion of space/digit/etc.
In some other places, ctype functions are used (or their wrappers), those can depend on the current locale setting.
It also looks like there is a subtle difference in ctype behavior on linux and macos (hence the tests failure).
It's still possible to use optimized ctype implementations in some cases; it's also possible to write functions that check that the current locale is never changed, so optimized routines can be used instead. Still, that's more complicated than I anticipated.

Let's make replacements step-by-step, without global changes. We don't have enough locale tests to make this kind of experiment right now.